+++ python/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Sarmata ASR gRPC Python client Changelog
 
+## [2.2.0] - 2018-12-12
+### Added
+- Support for setting gRPC deadline (how long the client is willing to wait for a reply from the server).
+
 ## [2.1.0] - 2018-04-13
 ### Added
 - Possibility to send service settings via request.